requirements.txt
absl-py==0.11.0
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
Cython==0.29.21
docker-pycreds==0.4.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
greenlet==0.4.17
grpcio==1.33.2
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
pathtools==0.1.2
Pillow==8.0.1
promise==2.3
protobuf==3.14.0
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
sentry-sdk==0.19.4
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.20
subprocess32==3.5.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.53.0
urllib3==1.26.2
watchdog==0.10.4
Werkzeug==1.0.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
